Extract Levenshtein: C
